Refactor Colaborador and Obra modules to use Integer as ID type; update service and repository methods for consistency

+++ src/main/java/br/com/softlutions/contrutora/modules/colaborador/repositories/ColaboradorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.softlutions.contrutora.modules.colaborador.entities.Colaborador;
 
-public interface ColaboradorRepository extends JpaRepository<Colaborador, Long> {
+public interface ColaboradorRepository extends JpaRepository<Colaborador, Integer> {
 }
